Added Game, MoveSequence and related types
