+++ Cargo.toml
@@ -7,3 +7,5 @@ edition = "2018"
 
 [dependencies]
 clap = "2.33.0"
+regex = "1.3.1"
+lazy_static = "1.4.0"
