+++ src/lib.rs
@@ -21,3 +21,5 @@
 #![warn(clippy::unimplemented, clippy::option_unwrap_used, clippy::result_unwrap_used)]
 
 pub mod chess;
+//pub mod engine;
+//pub mod protocol;
